requirements.txt
dj-database-url==0.4.1
Django==1.10
djangorestframework==3.4.6
gunicorn==0.14.2
Pillow==3.3.1
psycopg2==2.6.1
wheel==0.24.0
